% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-auxiliares.R
\name{adjacency2}
\alias{adjacency2}
\title{Processamento da matriz de adjacencia}
\usage{
${adjacency2}(${file,n.sites})
}
\arguments{
\item{file}{O nome do arquivo do qual os dados serão lidos.}

\item{n.sites}{Quantidade de unidades de area que a matriz de adjacencia possui}
}
\value{
uma matriz
}
\description{
Uma funcao que importa a matriz de adjacencia, formatando-a e deixando-a pronta para ser usada. Esta funcao foi projetada para importar a matriz
de adjacencia quando esta foi salva em formato .csv
}
\details{
The details of the adjacency2 function can be found in the vignette. Users can access the vignette using \verb{vignette(package = "SpatialBinReg")}.
}
\examples{

W=adjacency2('/home/alan/Documentos/TESE-ALAN/Dados/Pernambuco-trimestre-menos-chuvoso/W_PE.csv',n)

W_esparsa =W_sparsa(W) # calculando quantidades para a matriz de adjacencia esparsa W

diag(as.vector(W_esparsa$D_sparse)) # matriz diagonal com os vizinhos

}
\seealso{
\code{\link[base]{f}}
}
\author{
Alan Assunção
}
