% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-logito-R.R
\docType{data}
\name{Flogit}
\alias{Flogit}
\alias{F}
\title{Calculo da funcao de distribuicao acumulada da distribuicao Logistica (funcao de ligacao Logito)}
\format{
An object of class \code{logical} of length 1.
}
\usage{
${Flogit}(${x})
}
\arguments{
\item{x}{Um numero real}
}
\value{
um valor real entre zero e um.
}
\description{
Uma funcao que recebe dois valores, o valor de x e o valor do parametro de forma delta
    e calcula o valor da funcao de distribuicao acumulada para o modelo Logistico (funcao de ligacao Logito)
}
\details{
The details of the Flogit function can be found in the vignette. Users can access the vignette using \verb{vignette(package = "SpatialBinReg")}.
}
\examples{
Flogit(-0.5,2)

x=seq(-5,5,0.1)
delta <- 3

Flogit(x,delta)


}
\seealso{
\code{\link[base]{f}}
}
\author{
Alan Assunção
}
\keyword{datasets}
