% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-gumbel-reversa-reversa-de-potencia-R.R
\docType{data}
\name{Frprg}
\alias{Frprg}
\alias{F}
\title{Calculo da funcao de distribuicao acumulada da distribuicao Gumbel Reversa Reversa de Potencia}
\format{
An object of class \code{logical} of length 1.
}
\usage{
${Frprg}(${x, delta})
}
\arguments{
\item{x}{Um numero real}

\item{delta}{Um numero real}
}
\value{
um valor real entre zero e um.
}
\description{
Uma funcao que recebe dois valores, o valor de x e o valor do parametro de forma delta
    e calcula o valor da funcao de distribuicao acumulada para o modelo probabilistico Gumbel Reversa Reversa de Potencia
}
\details{
The details of the Frprg function can be found in the vignette. Users can access the vignette using \verb{vignette(package = "SpatialBinReg")}.
}
\examples{
Frprg(-0.5,2)

x=seq(-5,5,0.1)
delta <- 3

Frprg(x,delta)


}
\seealso{
\code{\link[base]{f}}
}
\author{
Alan Assunção
}
\keyword{datasets}
