% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-logito-R.R
\name{Glogit}
\alias{Glogit}
\title{Calculo da matriz de informacao de Fisher do modelo sob a funcao de ligacao  Logito}
\usage{
${Glogit}(${theta, y, X})
}
\arguments{
\item{theta}{Um vetor de parametros; a saber \beta (coeficientes da regressao) e os efeitos aleatorios \phi_{is} i=1,...,n e s=1,...,m}

\item{X}{Uma matriz de delineamento}

\item{Y}{Uma matriz de respostas binarias}
}
\value{
Uma matriz.
}
\description{
Uma funcao que recebe os valores dos parametros, a matriz de delineamento X e a matriz de respostas Y
    e calcula a matriz de informacao de Fisher para a funcao de ligacao  Logito
}
\details{
The details of the Glogit function can be found in the vignette. Users can access the vignette using \verb{vignette(package = "SpatialBinReg")}.
}
\examples{
beta1 = -0.7 ; beta2 = 0.7 # valores dos coeficientes

set.seed(1010)
   x1 = rnorm(50)
   x2 = rnorm(50)


   m=60 # quantidade de localizacoes
   X=cbind(x1,x2)
        
     betax=beta1*x1+beta2*x2

     
     Omega_sim=rgwish( n = 1, adj =W, b = m, D = S, threshold = 1e-8 ) # gerando a matriz de precisão
     
     Y=array(NA,c(50,m))
     phi=array(NA,c(50,m))
     
     for(k in 1:50)
     {
       phi[k,] = mvrnorm(1, mu = rep(0, times = m), Sigma = solve(Omega_sim))
       
       p = Fcloglog((betax[k]+phi[k,]))
       
       Y[k,] = rbinom(m, size=1, prob=p)
     }

     
     Gcloglog(theta=c(beta1,beta2,phi),y=Y,X=X)

}
\seealso{
\code{\link[base]{Fpc}}
}
\author{
Alan Assunção
}
