% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes-auxiliares.R
\name{W_sparsa}
\alias{W_sparsa}
\title{Calculo de varias quantidades relacionadas a matriz de adjacencia W}
\usage{
${W_sparsa}(${W})
}
\arguments{
\item{W}{Uma matriz}
}
\value{
um lista de  resulatdos, a saber: quantidade de unidades de area; quantidade de pares de adjacencia; pares de adjacencia; e numero de vizinhos
para cada unidade de area
}
\description{
Uma funcao que calcula varias quantidades relativas  a matriz de adjacencia W. Esta recebe a matriz de
adjacencia W, e retorna uma lista de resultados
}
\details{
The details of the W_sparsa function can be found in the vignette. Users can access the vignette using \verb{vignette(package = "SpatialBinReg")}.
}
\examples{
set.seed(1020304050)
 W1 = graph.sim( p = m, graph = "random", vis = FALSE,rewire = 0.05 ) # construIndo uma matriz de adjacencia
plot(W1)

W = matrix(W1,m,m) # matriz de adjacencia para estrutura grafica circular

W_esparsa =W_sparsa(W) # calculando quantidades para a matriz de adjacencia esparsa W

diag(as.vector(W_esparsa$D_sparse)) # matriz diagonal com os vizinhos

}
\seealso{
\code{\link[base]{f}}
}
\author{
Alan Assunção
}
